+++ modules-tutorial/tutorial1/src/main/java/org/openprovenance/prov/tutorial/tutorial1/Little.java
@@ -70,7 +70,7 @@ public class Little {
     public void doConversions(Document document, String file) {
         InteropFramework intF=new InteropFramework();
         intF.writeDocument(file, document);     
-        intF.writeDocument(System.out, Formats.ProvFormat.PROVN, document);
+        intF.writeDocument(System.out, document, Formats.ProvFormat.PROVN);
     }
 
     public void closingBanner() {